Implement a parameterized test attribute macro.
The macro is inspired by the parameterized tests annotation which can be found in JUnit.
An important difference to that macro, is that we are not limited to primitives in the attribute.
The attribute can take any valid expression.
Unlike JUnit, the current version does not support inputs defined by a function yet, however support for such a feature is planned.

The following example shows what a test with the paramaterized attribute could look like:

; #[parameterized(fruits = {
;   Fruit::Apple, Fruit::Pear, Fruit::Banana, Fruit::Bramble(BrambleFruit::Blackberry)
; }, names = {
;   "apple", "pear", "banana", "blackberry"
; })]

Note that the amount of arguments for the 'fruits' and 'names' parameters are the same (this is required for the generation of test cases).

The function signature could look like:

; fn a_fruity_test_case(fruits: Fruit, names: &str)

Here, we should note that the type of 'fruits' is the same as the type of the expressions in the attribute definition.
All inputs identified by 'fruits' should evaluate to an expression with the Fruit type.
Similarly, all inputs identified by 'names' should evaluate to expressions with the &str type.

If we assume the Fruit type has a method called 'name_of()' which outputs a &str representation of the Fruit, we could imagine
the body of the function to be:
; {
;   assert_eq!(fruits.name_of(), names);
; }

This complete example can be found in the 'examples/fruits.rs' folder. It can't be conveniently run from this place though,
so you should copy its contents to a place where you can easily run 'cargo test'.
Additional examples can be found in the 'tests/' folder or the 'expand' crate (this crate exists for debugging purposes).