Add transitive attributes support.
Use case:
(1) A test case can be marked with the #[should_panic] attribute, so
    a test would succeed if it panicked, and fail if it did not panic.

Currently however, we did not propagate (other) attributes on the parameterized
marked test function.
This commit ensures that other attributes placed on top of a parameterized
test function will also be included in all generated paramaterized test cases.