Accept an alternative workaround for the IDE test intent issue by allowing the #[test] attribute.
The test intent issue can be summarized as: "IDE's recognize #[test] attributes and provide useful
'run test' intents through their UI", but if test cases decorated by a #[test] attribute are generated
by a procedural macro, the IDE needs to be capable of expanding procedural macros." Most (all current?)
IDE's do not (yet) have this capability. Because these test intents can be practical during development
we attempt to provide workarounds.

The added workaround allows you to write a `#[test]` attribute after the `#[parameterized(...)]` attribute.
Because of parsing visibility, the parameterized attribute macro can't inspect any attribute defined
before itself (thus the ordering of attributes matters for this workaround!).

Two advantages of this approach over the original ide!() macro are: 1) we are not dependent on having an IDE
which is capable of expanding declarative macros, and 2) we don't need to stick our test cases into modules
as to only run the tests cases for one parameterized test function.