Remove removal of test macro attribute when defined after a parameterized attribute
Previously a user could specify a #[test] attribute after a #[parameterized(..)] attribute. We would then remove this #[test] attribute so the build would not fail.
It was originally added as an alternative to the ide!() macro / empty test case to force an IDE to show test module intents. The problem with this workaround is that the test intent on the module works (just like the ide! macro), but an additional test intent is created on the actual test (exactly what we would want (with the additional possible improvement where all generated tests would have an intent)!). The problem is that this last intent doesn't work (i.e. the test suit fails unexpectedly). This results in a more confusing situation than before where we simply didn't have this intent, therefore I've decided to remove it.