Release parameterized_macro 3.0.0
Strictly speaking, a major version bump was unnecessary. However
since it's parent crate `parameterized` depends on "2", it would
require a release itself to match the changed unstable feature, and
might cause issues for those who (too) strictly depend on
`parameterized 2.0.0`. Choosing a major version bump for the
macro crate was the safe choice.